Aula_Livre/requirements.txt
asgiref==3.11.0
Django==5.2.9
djangorestframework==3.16.1
sqlparse==0.5.4
tzdata==2025.2
